# back-end/src/modules/event-registrations/event-registrations.controller.ts
import { Body, Controller, Delete, Get, Param, ParseIntPipe, Post, Query } from '@nestjs/common';
import { ApiTags } from '@nestjs/swagger';
import { Roles } from '../../common/decorators/roles.decorator';
import { AppRole } from '../rbac/role.enum';
import { EventRegistrationsService } from './event-registrations.service';

@ApiTags('event-registrations')
@Controller('event-registrations')
export class EventRegistrationsController {
  constructor(private readonly service: EventRegistrationsService) {}

  @Get()
  @Roles(AppRole.ADMIN, AppRole.COMMUNITY_MANAGER, AppRole.MODERATOR, AppRole.USER, AppRole.ORGANIZER)
  findAll(@Query('userId') userId?: string, @Query('eventId') eventId?: string) {
    return this.service.findAll(userId ? Number(userId) : undefined, eventId ? Number(eventId) : undefined);
  }

  @Post()
  @Roles(AppRole.ADMIN, AppRole.COMMUNITY_MANAGER, AppRole.MODERATOR, AppRole.USER, AppRole.ORGANIZER)
  create(@Body() payload: any) {
    return this.service.create(payload);
  }

  @Delete(':id')
  @Roles(AppRole.ADMIN, AppRole.COMMUNITY_MANAGER, AppRole.MODERATOR, AppRole.USER, AppRole.ORGANIZER)
  remove(@Param('id', ParseIntPipe) id: number) {
    return this.service.remove(id);
  }
}